refactor(frontend): use mappers index import
