+++ frontend/__tests__/.server/domain/mappers/federal-government-insurance-plan.dto.mapper.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import type { FederalGovernmentInsurancePlanDto } from '~/.server/domain/dtos/federal-government-insurance-plan.dto';
 import type { FederalGovernmentInsurancePlanEntity } from '~/.server/domain/entities/federal-government-insurance-plan.entity';
-import { FederalGovernmentInsurancePlanDtoMapperImpl } from '~/.server/domain/mappers/federal-government-insurance-plan.dto.mapper';
+import { FederalGovernmentInsurancePlanDtoMapperImpl } from '~/.server/domain/mappers';
 
 describe('FederalGovernmentInsurancePlanDtoMapperImpl', () => {
   afterEach(() => {
