+++ frontend/__tests__/.server/domain/mappers/preferred-language.dto.mapper.test.ts
@@ -3,7 +3,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { PreferredLanguageDto } from '~/.server/domain/dtos/preferred-language.dto';
 import type { PreferredLanguageEntity } from '~/.server/domain/entities/preferred-language.entity';
-import { PreferredLanguageDtoMapperImpl } from '~/.server/domain/mappers/preferred-language.dto.mapper';
+import { PreferredLanguageDtoMapperImpl } from '~/.server/domain/mappers';
 
 describe('PreferredLanguageDtoMapperImpl', () => {
   afterEach(() => {
