+++ frontend/__tests__/.server/domain/mappers/province-territory-state.dto.mapper.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
 import type { ProvinceTerritoryStateEntity } from '~/.server/domain/entities/province-territory-state.entity';
-import { ProvinceTerritoryStateDtoMapperImpl } from '~/.server/domain/mappers/province-territory-state.dto.mapper';
+import { ProvinceTerritoryStateDtoMapperImpl } from '~/.server/domain/mappers';
 
 describe('ProvinceTerritoryStateDtoMapperImpl', () => {
   afterEach(() => {
