+++ frontend/__tests__/.server/domain/mappers/provincial-government-insurance-plan.dto.mapper.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import type { ProvincialGovernmentInsurancePlanDto } from '~/.server/domain/dtos/provincial-government-insurance-plan.dto';
 import type { ProvincialGovernmentInsurancePlanEntity } from '~/.server/domain/entities/provincial-government-insurance-plan.entity';
-import { ProvincialGovernmentInsurancePlanDtoMapperImpl } from '~/.server/domain/mappers/provincial-government-insurance-plan.dto.mapper';
+import { ProvincialGovernmentInsurancePlanDtoMapperImpl } from '~/.server/domain/mappers';
 
 describe('ProvincialGovernmentInsurancePlanDtoMapperImpl', () => {
   afterEach(() => {
