+++ frontend/__tests__/.server/domain/services/client-friendly-status.service.test.ts
@@ -3,7 +3,7 @@ import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos/client-friendly-status.dto';
-import type { ClientFriendlyStatusDtoMapper } from '~/.server/domain/mappers/client-friendly-status.dto.mapper';
+import type { ClientFriendlyStatusDtoMapper } from '~/.server/domain/mappers';
 import type { ClientFriendlyStatusRepository } from '~/.server/domain/repositories/client-friendly-status.repository';
 import { ClientFriendlyStatusServiceImpl } from '~/.server/domain/services/client-friendly-status.service';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
