+++ frontend/__tests__/.server/domain/services/country.service.test.ts
@@ -3,7 +3,7 @@ import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { CountryDto } from '~/.server/domain/dtos/country.dto';
-import type { CountryDtoMapper } from '~/.server/domain/mappers/country.dto.mapper';
+import type { CountryDtoMapper } from '~/.server/domain/mappers';
 import type { CountryRepository } from '~/.server/domain/repositories/country.repository';
 import { CountryServiceImpl } from '~/.server/domain/services/country.service';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
