+++ frontend/__tests__/.server/domain/services/preferred-language.service.test.ts
@@ -3,7 +3,7 @@ import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { PreferredLanguageDto } from '~/.server/domain/dtos/preferred-language.dto';
-import type { PreferredLanguageDtoMapper } from '~/.server/domain/mappers/preferred-language.dto.mapper';
+import type { PreferredLanguageDtoMapper } from '~/.server/domain/mappers';
 import type { PreferredLanguageRepository } from '~/.server/domain/repositories/preferred-language.repository';
 import { PreferredLanguageServiceImpl } from '~/.server/domain/services/preferred-language.service';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
