+++ frontend/__tests__/.server/domain/services/province-territory-state.service.test.ts
@@ -3,7 +3,7 @@ import { mock } from 'vitest-mock-extended';
 
 import type { ServerConfig } from '~/.server/configs/server.config';
 import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
-import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers/province-territory-state.dto.mapper';
+import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers';
 import type { ProvinceTerritoryStateRepository } from '~/.server/domain/repositories/province-territory-state.repository';
 import { ProvinceTerritoryStateServiceImpl } from '~/.server/domain/services/province-territory-state.service';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
