+++ frontend/app/.server/domain/services/marital-status.service.ts
@@ -4,7 +4,7 @@ import moize from 'moize';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { MaritalStatusDto } from '~/.server/domain/dtos/marital-status.dto';
-import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers/marital-status.dto.mapper';
+import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers';
 import type { MaritalStatusRepository } from '~/.server/domain/repositories/marital-status.repository';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
