+++ frontend/app/.server/domain/services/province-territory-state.service.ts
@@ -4,7 +4,7 @@ import moize from 'moize';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { ProvinceTerritoryStateDto } from '~/.server/domain/dtos/province-territory-state.dto';
-import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers/province-territory-state.dto.mapper';
+import type { ProvinceTerritoryStateDtoMapper } from '~/.server/domain/mappers';
 import type { ProvinceTerritoryStateRepository } from '~/.server/domain/repositories/province-territory-state.repository';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
